SourceCode/app/agents/graph/stage.py
from typing import Annotated
from langgraph.graph.message import add_messages
from typing_extensions import TypedDict

class AgentState(TypedDict):
    messages: Annotated[list, add_messages]
    risk_level: str
    query_type: str