SourceCode/pyproject.toml
[project]
name = "sourcecode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.68.0",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "openai>=2.29.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
